feat: removed deprecated facet search resources, added contingency service and table resource #3893
